Move mPRM fixtures to own file
