+++ tests/mocks/mock_websocket.py
@@ -1,0 +1,3 @@
+class MockWebsocket:
+    def close(self):
+        pass
