+++ tests/test_mprm_rest.py
@@ -1,6 +1,7 @@
 import pytest
+from requests import ConnectTimeout
 
-from devolo_home_control_api.backend.mprm_rest import MprmRest, MprmDeviceCommunicationError
+from devolo_home_control_api.backend.mprm_rest import MprmDeviceCommunicationError, MprmRest
 from devolo_home_control_api.mydevolo import Mydevolo
 
 
@@ -8,9 +9,7 @@ from devolo_home_control_api.mydevolo import Mydevolo
 @pytest.mark.usefixtures("mock_mydevolo__call")
 class TestMprmRest:
 
-    @pytest.mark.usefixtures("mock_mprmrest__extract_data_from_element_uid")
-    @pytest.mark.usefixtures("mock_mprmrest__post")
-    def test_get_name_and_element_uids(self):
+    def test_get_name_and_element_uids(self, mock_mprmrest__extract_data_from_element_uid, mock_mprmrest__post):
         properties = self.mprm.get_name_and_element_uids("test")
         assert properties == {"itemName": "test_name",
                               "zone": "test_zone",
@@ -34,11 +33,11 @@ class TestMprmRest:
         second = MprmRest.get_instance()
         assert first is second
 
-    def test_create_connection_local(self, mock_get_local_session):
+    def test_create_connection_local(self, mock_mprmrest_get_local_session):
         self.mprm._local_ip = "123.456.789.123"
         self.mprm.create_connection()
 
-    def test_create_connection_remote(self, mock_get_remote_session, mydevolo):
+    def test_create_connection_remote(self, mock_mprmrest_get_remote_session, mydevolo):
         self.mprm._gateway.external_access = True
         self._mydevolo = Mydevolo.get_instance()
         self.mprm.create_connection()
@@ -48,34 +47,32 @@ class TestMprmRest:
             self.mprm._gateway.external_access = False
             self.mprm.create_connection()
 
-    @pytest.mark.usefixtures("mock_mprmrest__post")
-    def test_extract_data_from_element_uid(self):
+    def test_extract_data_from_element_uid(self, mock_mprmrest__post):
         properties = self.mprm.extract_data_from_element_uid(uid="test")
         assert properties.get("properties").get("itemName") == "test_name"
 
-    @pytest.mark.usefixtures("mock_mprmrest__post")
-    def test_get_all_devices(self):
+    def test_get_all_devices(self, mock_mprmrest__post):
         devices = self.mprm.get_all_devices()
         assert devices == "deviceUIDs"
 
     @pytest.mark.usefixtures("mock_session_post")
     @pytest.mark.usefixtures("mock_response_json")
     def test_get_local_session_valid(self):
+        # TODO: Move IP to test_data
         self.mprm._local_ip = "123.456.789.123"
         self.mprm.get_local_session()
 
     @pytest.mark.usefixtures("mock_response_json_ConnectTimeout")
     def test_get_local_session_ConnectTimeout(self):
-        from requests import ConnectTimeout
+        # TODO: Move IP to test_data
         self.mprm._local_ip = "123.456.789.123"
-
         with pytest.raises(ConnectTimeout):
             self.mprm.get_local_session()
 
     @pytest.mark.usefixtures("mock_response_json_JSONDecodeError")
     def test_get_local_session_JSONDecodeError(self):
+        # TODO: Move IP to test_data
         self.mprm._local_ip = "123.456.789.123"
-
         with pytest.raises(MprmDeviceCommunicationError):
             self.mprm.get_local_session()
 
@@ -102,7 +99,6 @@ class TestMprmRest:
         with pytest.raises(ValueError):
             self.mprm.post({"data": "test"})
 
-    @pytest.mark.usefixtures("mock_response_requests_valid")
-    def test_post_valid(self):
+    def test_post_valid(self, mock_response_requests_valid):
         self.mprm._data_id = 1
         assert self.mprm.post({"data": "test"}).get("id") == 2
